rename {RandomGenerator => random_generator}.hpp
